+++ Assets/_Laz/Scripts/Lazo/FakeLazo.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b3f5d949c92f6144d9c10be572f7b38b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
