+++ Assets/_Laz/Scripts/Lazo/FakeLazo/FakeLazo.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bf0bed270acbd5945be5be6df2f46051
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
