+++ Assets/_Laz/Scripts/Lazo/FakeLazo/FakeLazoBehaviour.cs
@@ -1,6 +1,3 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using PerigonGames;
 using Sirenix.OdinInspector;
 using UnityEngine;
@@ -17,7 +14,6 @@ namespace Laz
         
         public void SetLine(FakeLazo fakeLazo)
         {
-            // Set the color as blue
             SetLineFrozen();
             HandleOnListOfLazoPositionsChanged(fakeLazo.Positions);
             _fakeLazo = fakeLazo;
@@ -71,38 +67,4 @@ namespace Laz
             CleanUp();
         }
     }
-
-    public class FakeLazo
-    {
-        private List<LazoPosition> _listOfPositions;
-        
-        public event Action<Vector3[]> OnListOfLazoPositionsChanged;
-        public Vector3[] Positions { get; }
-
-        public FakeLazo(List<LazoPosition> listOfPoints)
-        {
-            _listOfPositions = listOfPoints;
-            Positions = _listOfPositions.Select(point => point.Position).ToArray();
-        }
-        
-        private void RemoveOldestPointIfNeeded(float deltaTime)
-        {
-            _listOfPositions = _listOfPositions.Select(lazoPosition =>
-            {
-                lazoPosition.DecrementTimeToLiveBy(deltaTime);
-                return lazoPosition;
-            }).Where(lazoPosition => !lazoPosition.IsTimeBelowZero).ToList();
-            OnLazoPositionsChanged();
-        }
-        
-        private void OnLazoPositionsChanged()
-        {
-            if (OnListOfLazoPositionsChanged != null)
-            {
-                OnListOfLazoPositionsChanged(_listOfPositions.Select(x => x.Position).Reverse().ToArray());
-            }
-        }
-    }
-    
-    
 }
